Poetry-Analyzer/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
langdetect
torch
numpy
pydantic
